making things doubles, breaking shit.

+++ Quad/Controller.cpp
@@ -5,7 +5,7 @@ PID rollPID(1, .1, .5); //TODO config this
 PID yawPID(1, 1, 1);
 PID vertVelocityPID(1, .5, .5);
 
-void Controller::calcPryCorrection(float pitch, float roll, float yaw,
+void Controller::calcPryCorrection(double pitch, double roll, double yaw,
                                    int* pitchCorrection, int* rollCorrection, int* yawCorrection){
     *pitchCorrection = pitchPID.compute(pitch);
     *rollCorrection = 0;
@@ -16,7 +16,7 @@ void Controller::calcPryCorrection(float pitch, float roll, float yaw,
  * Sets the desired orientation of the quad through
  *  the pitch roll and yaw values in degrees
  */
-void Controller::setDesiredPry(float pitch, float roll, float yaw){
+void Controller::setDesiredPry(double pitch, double roll, double yaw){
   pitchPID.setDesired(pitch);
   rollPID.setDesired(roll);
   yawPID.setDesired(yaw);
